+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
-## [0.1.3] - 2023-02-28
+## [0.1.3] - 2023-03-01
 
 ### Added in 0.1.3
 
-- GetObservers()
+- `subject.GetObservers()`
+- `notifier.Notify()`
 
 ## [0.1.2] - 2023-02-23